displayException: Remove trailing newline (#37)
* displayException: Remove trailing newline

In #36, I restructured the `AnnotatedException` `displayException`
implementation so that instead of being of the form:

    displayException =
        unlines [ ... ] <> callStackMessage
        where
            callStackMessage = prettyCallStack frames

It was of the form:

    displayException =
        unlines $ [ ... ] <> callStackMessage
        where
            callStackMessage = [prettyCallStack frames]

I didn't realize that `unlines` adds a trailing newline, which is
generally the wrong behavior for `displayException` instances.

This change replaces `unlines` with `concat . intersperse "\n"` in order
to regain the correct behavior.

* 0.3.0.3 -> 0.3.0.4